Migrate code from experimental repository
This commit includes the initial implementation, extracted from an
experimental Rails application used as a test bed for the initial
interface.

The changes to the `README.md` provide a high level explanation on how
to consume the gem.

From an architectural perspective, the project is highly inspired by the
[`ActionMailer` portion of Rails][actionmailer].

The `ActionClient::Base` class descends from
[`AbstractController::Base`][abstract_controller] so that it provides
consumable "actions" and its ActionView-aware `render` method.

There are also a collection of internal routes and controllers the gem
provides in order to render its own set of
[`ActionMailer::Preview`-inspired][preview] `ActionClient::Preview`
classes.

[actionmailer]: https://github.com/rails/rails/blob/master/actionmailer/README.rdoc
[abstract_controller]: https://api.rubyonrails.org/classes/AbstractController/Base.html
[preview]: https://guides.rubyonrails.org/action_mailer_basics.html#previewing-emails