README tweaks (#16)
* Fix typo in README example.

* Change default headers example in README.

ActionClient will set the `Content-Type` header automatically based on the
request body's template type. As a result, an example of default headers
that sets `Content-Type` could be a little misleading.

This commit replaces it with an authorization example, which is likely to be
a common reason to set default headers.