Add test coverage for HttpStatusFilter
Prior to this commit, any calls to `after_submit` with an [Endless
Range][endless] would result in an error, since the naive implementation
would coerce all values to an `Array`.

This commit adds test coverage to ensure that existing behavior remains
intact, and then extends the `HttpStatusFilter` to support Endless
Ranges.

[endless]: https://ruby-doc.org/core-2.7.1/Range.html#class-Range-label-Beginless-2FEndless+Ranges