Make after_submit use ActiveSupport::Callbacks
Replace the `after_submit` callback system's middleware backend with
[`ActiveSupport::Callbacks`][Callbacks].

The change is mostly an implementation detail, and still supports
`after_submit` blocks with 0, 1, or 3 arguments.

The biggest impact to consumer applications is that the `after_submit`
is not required to return a value. This means that re-writing the
response body requires an assignment to `response.body`.

The blocks are evaluated in the context of the `ActionClient::Base`
instance, making the `response` available to the scope.

[Callbacks]: https://api.rubyonrails.org/classes/ActiveSupport/Callbacks.html