CI: Install Rails version based on Gemfile
Prior to this commit, our Continuous Integration test suite was not
being run against differing versions of Rails.

This commit resolves that by passing the `rails_constraint` to the `gem
"rails"` declaration, which we determine based on the value of the
`RAILS_VERSION` environment variable.

Additionally, for the sake of backwards compatibility with `rails@5.2`,
this commit also includes implementation fixes necessary to get the Test
Suite passing.