Introduce a Rails generator
This commit introduces the action_client generator (started with the
rails [generator generator]).

It is inspired by the Action Mailer generators, including the [Mailer
generator], [Test and preview generator], and [View generator].

The generator can be run within a Rails project with

```
bin/rails generate action_client articles create show
```

This will create the ArticlesClient class, as well as the view directory
with a keep file, a config file, and an ArticlesClientPreview for
testing your client.

The client name can be passed CamelCased or under_scored, with or
without the client prefix: "articles", "Articles", "articles_client",
or "ArticlesClient".

The client's actions may be passed as additional arguments. These will
be used to define methods in the client and preview classes.

[generator generator]: https://edgeguides.rubyonrails.org/generators.html#creating-generators-with-generators
[Mailer generator]: https://github.com/rails/rails/tree/master/actionmailer/lib/rails/generators/mailer
[Test and preview generator]: https://github.com/rails/rails/tree/master/railties/lib/rails/generators/test_unit/mailer
[View generator]: https://github.com/rails/rails/tree/master/railties/lib/rails/generators/erb/mailer