Declare Constants in test suite
The majority of the test suite's coverage does not rely on a
`ActionClient::Base` descendant or an `ActionClient::SubmissionJob`
descendant being declared as a constant.

For others, the constant itself is unimportant, but the Client's
[`controller_path` method][controller_path] reflexively resolves the
path based on the class name. For example, `ArticleClient` will resolve
to `article_client`.

For some tests, the constant _must_ exist. For example, when serializing
the request to the `ActionClient::SubmissionJob`, the classes cannot be
anonymously declared.

To ensure those tests can function, this commit restructures the
`declare_client` helper to set and unset constants, ensuring they're
unset between tests. Additionally, add the `declare_job` helper to
declare `ActionClient::SubmissionJob` descendants.

[controller_path]: https://api.rubyonrails.org/classes/AbstractController/Base.html#method-c-controller_path